Fix: Create shared WebSocket types package for frontend/backend synchronization
